+++ jsk_pcl_ros/src/edgebased_cube_finder_nodelet.cpp
@@ -1206,6 +1206,5 @@ namespace jsk_pcl_ros
 }
 
 #include <pluginlib/class_list_macros.h>
-typedef jsk_pcl_ros::EdgebasedCubeFinder EdgebasedCubeFinder;
-PLUGINLIB_DECLARE_CLASS (jsk_pcl, EdgebasedCubeFinder, EdgebasedCubeFinder, nodelet::Nodelet);
+PLUGINLIB_EXPORT_CLASS (jsk_pcl_ros::EdgebasedCubeFinder, nodelet::Nodelet);