+++ jsk_pcl_ros/src/environment_plane_modeling_nodelet.cpp
@@ -963,5 +963,5 @@ namespace jsk_pcl_ros
   
 }
 
-typedef jsk_pcl_ros::EnvironmentPlaneModeling EnvironmentPlaneModeling;
-PLUGINLIB_DECLARE_CLASS (jsk_pcl, EnvironmentPlaneModeling, EnvironmentPlaneModeling, nodelet::Nodelet);
+PLUGINLIB_EXPORT_CLASS (jsk_pcl_ros::EnvironmentPlaneModeling,
+                        nodelet::Nodelet);