+++ jsk_pcl_ros/src/organized_edge_detector_nodelet.cpp
@@ -357,5 +357,4 @@ namespace jsk_pcl_ros
 }
 
 #include <pluginlib/class_list_macros.h>
-typedef jsk_pcl_ros::OrganizedEdgeDetector OrganizedEdgeDetector;
-PLUGINLIB_DECLARE_CLASS (jsk_pcl, OrganizedEdgeDetector, OrganizedEdgeDetector, nodelet::Nodelet);
+PLUGINLIB_EXPORT_CLASS (jsk_pcl_ros::OrganizedEdgeDetector, nodelet::Nodelet);